Move Book into a class
